requirements.txt
typer>=0.9.0
rich>=13.0.0
demucs>=4.0.0
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1
ffmpeg-python>=0.2.0
numpy>=1.24
scipy>=1.10
